rawdata should no longer be dropped from output file
added a check to reinitialize an existing channel if it now contains data. Modified Channel{T} to maintain already defined props.

also have a function readseginfo that is somewhat a WIP that will read only the metadata and segment information for an idea of what is contained in the file.